Rewrite compressor and decompressor APIs
The old API was what I managed to hack up in a first pass. It was never
intended to live long.

This commit introduces new APIs for compressio and decompression. We
introduce ZstdCompressor and ZstdDecompressor classes. These allow you
to compress and decompress data N ways.

A benefit of this API is it separates compression and decompression
parameters from the mechanism by which compression and decompression
occur. It also allows the reuse of compression and decompression
contexts without having to instantiate a new Python object.

Tests and docs have been updated.