Use a dedicated type to represent compression dictionary data
Before, we just shuffled bytes around. This was a bit too low-level
for my liking, especially since the Zstandard APIs around dictionaries
are marked as experimental.

This commit introduces the ZstdCompressionDict type to represent
dictionaries and abstract Python away from the C API.

The low-level "dictionary_id()" function has been moved to a method
of this new type.