Rewrite ZstdDecompressorIterator_iternext
As I said in the commit message that introduced this code, I wasn't too
thrilled about the state of the code. As I was writting fuzzing tests,
I found a case where the code completely failed. I also managed to cause
heap corruption with the old code as part of testing on Linux.

The new code is cleaner and hopefully doesn't have the bugs of the older
code.