Move buffer sizes to layer below compressor and decompressor
Before, the compressor and decompressor structs held static
input and output buffer sizes. In preparation for making buffer
sizes configurable, we move these values to the structs or
functions holding state for individual operations. I could have
retained the values on the compressor and decompressor instances.
However, I feel buffer sizes are a per-operation value as opposed
to per-compressor/decompressor. This does result in more function
calls to the zstd C API to obtain the buffer sizes. However, that
function should be cheap compared to the actual compression
and decompression functions.