Don't assume frame is finished if ZSTD_decompressStream returns 1
The docs in zstd.h for version 1.0 say ZSTD_decompressStream will
return 1 if there is data left in an internal buffer. These docs
were misleading and led us to believe that the frame was fully read
and there was nothing else to read. The docs have been updated
upstream to indicate that a return of 1 means there is still
work to do.

So, remove the branch checking for ZSTD_decompressStream returning 1
and treat all positive values as "call again."

This fixes #1.