Add arguments to control read and write sizes
As part of testing this, it was discovered that the logic in the
read_from compressor iterator was a bit flawed. Specifically, there
were a few incorrect assertions regarding boundary conditions. The
logic to feed data into the compressor and end the stream was also
a bit repetitive. And, we would continue to read from the source if
we knew it was exhausted. There was also a memory leak if the input
buffer was not fully consumed. These issues have all been addressed.

Also, a few instances of ZSTD_decompressStream returning 1 being
interpretted as "finished" (similar to issue #2) were also addressed.