CI support for Conda packages
Travis and AppVeyor now build Conda packages for 64-bit Python
2.7 and 3.5.

Because Conda packages aren't using Tox, we now have scripts in
place that take different code paths depending on whether we're
performing a Conda build.

Tweaks were made to the previously unused code in
appveyor-bootstrap.ps1 to download the latest versions of
Miniconda and to update the install.