Add "compressobj" API for compatibility with Python standard library
As part of teaching Mercurial to use zstandard, I encountered an
abstraction where Mercurial used the common "compress()" and "flush()"
methods to abstract over compression algorithms. Since this API
is common in the Python standard library, it makes sense to expose
it so consumers can code to a standard interface. This commit does
that.