Support streaming decompression of objects conforming to buffer protocol
Before, we required input objects to expose a .read(size) method.
Oftentimes we want to use a simple data type, such as bytes or
a memoryview. We teach ZstdDecompressor.read_from() to support
reading from objects that expose the buffer protocol and make
their data available as a contiguous range of bytes.