Fix installation of test files in Conda
A version of a Conda package rolled out in the past few months broke
`conda build` when processing test files. It has something to do with
conda_build.create_test and conda_build.utils.copy_into(). It looks
like that copy_into() function isn't ensuring the destination directory
exists. I think this is fixed upstream. In the meantime, a workaround
is to list the directory instead of using a wildcard.