Upgrade to zstd 1.1.2
Sources imported without modifications.

The zbuff library has been deprecated as of this release. The files
are now only referenced/built if legacy support is needed.

zstd.h changed the ZSTDLIB_API #define in a way that confused the CFFI
parser. So we had to hack make_cffi.py to strip the code it can't read.