Feed preprocessed output into CFFI
Before, we hacked zstd.h so CFFI could parse it. This was a bit too...
hacky.

In this commit, we instead invoke the C preprocessor on zstd.h and feed
that output into CFFI. This resolves all the complex expressions that
CFFI's preprocessor can't yet parse and appears to "just work."

As a side effect, constants should be more accurate in 32 bit builds
since they are actually evaluated now. And, we're better protected
from changes in the upstream source, making upgrades easier.