Always execute vcvarsall.bat in AppVeyor
Previously, some Python versions were relying on functionality in
modern distutils to find the appropriate toolchain. This kinda/sorta
worked, but was inconsistent and turned out to be a bit fragile.
For example, it was executing a script in the SDK directory, which
hard coded certain compiler versions.

In this commit, we change our script that sets up the environment
to always call vcvarsall.bat to set up the environment. We also
tell distutils to always use the settings exported by the
environment instead of guessing them. The execution environment
is consistent between Python versions and no longer relies on
smarts in distutils.