cffi: don't drop references to allocated objects
According to CFFI's docs, CFFI is very sensitive about retaining
references to the original object return by ffi.new(). For
example, assigning the result of ffi.new() to a struct member
will convert the value and drop a reference to the original
item, possibly freeing it. This could lead to access to corrupt
member.

So, this commit introduces tons of intermediary variables to prevent
that from happening.