compressor: support multi-threaded compression APIs
The "threads" argument has been added to ZstdCompressor. It allows
controlling the number of threads to use for compression. When
specified, a ZSTDMT_CCtx instead of ZSTD_CCtx is constructed.

Multi-threaded compression requires the ZSTDMT_* APIs, which are
distinct from the non-multithreaded APIs. This means existing code needs
to be taught to call these functions. This commit does that for
compress() and copy_stream(). Other methods raise NotImplementedError
for now.

Basic tests have been added. There is still a bunch more to do. But we
have to start somewhere.