compressor: allocate multiple buffers in batch worker
Previously, we allocated a single buffer and continuously
reallocated it to a larger size to hold the compressed frames.

In this commit, we change the logic to allocate multiple buffers
per worker.

The size of the buffers is 1:16 of the input size. For most inputs,
this will result in <10 buffers. That's more than 1 but not a lot.
It seems a reasonable default.

As the inline comments say, there is still room to improve the
behavior. But this feels strictly better than what existed
before.