compressor: add explicit "threads" argument to multi_compress_to_buffer()
"threads" on ZstdCompressor() is used to engage the multi-threaded zstd
APIs. This behavior is logically different than using N threads to process
a batch of work. There was no such "threads" argument on ZstdDecompressor(),
but there was on multi_decompress_to_buffer(). So adding an explicit
"threads" argument to multi_compress_to_buffer() makes behavior
consistent.