global: provide a zstandard Python module
Previously, we required users to import either "zstd" or
"zstd_cffi" depending on whether they wanted the C or CFFI
backend, respectively. This was annoying. Plus, "zstd"
didn't match the PyPI package name ("zstandard"). And,
the "zstd" PyPI package also used the "zstd" module.

On top of that, not having a common Python module meant that
all functionality needed to be implemented in both the C
and CFFI backends. This was acceptable before because we
didn't really have any functionality that needed implemented
in Python. But, at least one upcoming feature (file object
API) will be easier to implement with it.

So, we introduce a "zstandard" Python package that imports
and re-exports a backend. See the inline comments for
why this is more complicated than the trivial implementation.