tests: import and use zstandard module for cffi test variation
This required some module import hackiness. Before, we relied on
the backend modules being different. Now, we need to import a
different version of the module. But if we just did an `import`
it would essentially no-op because the module is already in
`sys.modules`. So, we invoke lower-level module importing
functionality to load the source file into a module of a
different name. Then we do as we did before and replace the
"zstd" symbol in the copy of the class/methods we create.

In addition, this adds some logic that verifies the test module
has a `zstd` symbol and skips the cffi variation if the imported
backend is already cffi. This prevents redundant running when
the C extension is not available.