tests: limit window log size on 32-bit machines
zstd 1.1.4 increased WINDOWLOG_MAX on 32-bit machines. CI is having
trouble allocating a continuous memory region the size of the new
max. So, change the test to cap the max window log on 32-bit Pythons.

This does sacrifice some bounds checking when fuzz testing. If there
is a bug here, 64-bit should find it.