setup: drop support for Python 2.6
Python 2.6 is a dead version of Python. The main reason I
made python-zstandard work with 2.6 is because I wanted
python-zstandard to be usable by Mercurial, which still
supported 2.6. Well, the next release of Mercurial (4.3) is dropping
Python 2.6 support. In the absence of another high visibility
consumer, I've decided to drop support for Python 2.6.