frameparams: use buffer protocol in get_frame_parameters (#26)
This is slightly more involved than other changes because of
existing checking for unicode types. Python 2 happily allows
unicode to be cast to a Py_buffer. Python 3 doesn't. So we
need to update the test accordingly.

The CFFI code has also been updated to use ffi.from_buffer()
so it is consistent with the C code.

And since this is the last function using y#/s# for argument
parsing, update the NEWS.