setup: filter out private functions from zstdmt_compress.h
The presence of these functions confuses cffi. Specifically, they are
referencing incomplete types (e.g. ZSTD_CCtx). I think because the
functions are not exported, the compiler insists that the types be
complete at compile time.

Since these are private functions and we don't need them for
python-zstandard, we simply filter them away when doing the cffi
processing.

With this change, cffi compiles cleanly after the 1.3.2 update.