compressor: switch to new advanced API
There is a new C API for managing individual compression parameters
and a new API for advanced compression. IMO these APIs are superior
to the previous APIs. But the real reason to adopt the new APIs
is that you can do things with them that you can't with the old APIs.
For example, you can use the new long distance matching mode.

This (massive) commit rips out code using the old APIs and replaces it
with use of ZSTD_CCtx_params, ZSTD_compress_generic(),
ZSTD_CCtx_setPledgedSrcSize() and friends.

Test changes are surprisingly minimal.