compressionparams: revamp to hold all compression parameters
There are a ton of parameters to control compression settings.
With us now using the new advanced and generic API for compression
where these parameters are tracked on a dedicated compression
parameters instance, it makes sense for our existing CompressionParameters
type to gain the ability to hold these new parameters.

This commit teaches CompressionParameters to define all compression
parameter settings.

As part of this commit, legacy APIs have been removed. Behavior of
ZstdCompressor.__init__ has also changed with regards to which
arguments are allowed when.

The goal is to expose a minimal set of high-level parameters on
ZstdCompressor and require the use of CompressionParameters for
anything more advanced.