tests: update CompressionParameters fuzzing test
Now that there is a dedicated compression parameters type in zstd,
we no longer need to instantiate a compressor to validate the
parameters. So we delete that code.

Since we no longer create a compressor, we no longer have to worry
about excessive memory allocation. So we can remove code that
existed to prevent us from hitting memory limits.

This deleted code appeared to also be responsible for an Hypothesis
warning about run times. I suspect the allocation of 1+ GB of
memory was taking a few hundred milliseconds and Hypothesis
complained about that.

While I was here, I also double checked whether ZSTD_checkCParams()
is moving the goal posts on search length. It is not. So that
workaround was also deleted.