cffi: report sizes of objects to aid garbage collection
Running bench.py with PyPy reveals an apparant memory leak in
some tight loops. On further investigation, this doesn't appear
to actually be a memory leak but rather PyPy's garbage collection
heuristics not being as aggressive as CPython's.

CFFI 1.11 supports declaring the size of an object when registering
a garbage collection function. This metadata is used by PyPy to
trigger GC sooner.

This commit defines sizes of ZSTD_CCtx, ZSTD_DCtx, ZSTD_CDict, and
ZSTD_DDict instances. By doing so, PyPy still "leaks" memory. But
the volume of the leaking is greatly reduced. So I think this has the
desired impact.

FWIW, the leak in bench.py for single use ZstdCompressor is still in
the ~5 GB range for ~400,000 chunks. That's better than ~8 GB. But it
is still quite absurd IMO. It is worth digging into this issue more.