global: validate Py_buffer instances are contiguous
I'm not sure if it is possible with any CPython types, but
Py_buffer instances can be backed by non-contiguous memory. Or
the buffer pointer may point to inside the buffer or the end.

You need to validate that the buffer is contiguous before you
can safely access it.

This commit adds checks for everywhere we accept a Py_buffer that
ensure the underlying memory is C contiguous and has at most 1
dimension.

In the real world, sometimes buffers will have 0 dimensions. But
types like bytes are actually 1 dimension, which each element being
of length 1. So we need to allow 1 dimension instances as well.

I can't think of a case where we'd need to allow more than 1
dimension. So the cap is 1.