compressor: write content sizes into frame headers by default
This is what zstandard does. We should follow their lead.

A reason this wasn't the default before is because we didn't
have an API for disambiguating between a frame size of 0 and
no content size in the frame. So we couldn't round trip empty
frames with compress()/decompress(), which was kind of bad.