ci: use default Python install in Appveyor
Attempting a custom install of Python was a noble goal in order
to achieve determinism over time. But it was fragile and things
broke if our installed version didn't match Appveyor's version
exactly.

Let's give up and use the Python provided by Appveyor.

We could probably delete some Powershell code after this. But
there's little harm in keeping it around.