decompressobj: ensure all data is emitted from decompress()
I managed to trigger this bug in Mercurial. Not sure of a good
way to test it explicitly beyond the existing fuzzing tests. I
think a problem is the output buffer is always large and none of
our input is large enough to fill the output buffer and trigger
the buggy behavior. But using this code with Mercurial fixes the
problem.

I also opened facebook/zstd/issues/1360 to track potentially
improving the zstd docs.