c-ext: track compression parameters on ZSTD_CCtx_params instances
Now that we have an API for retrieving parameters on
ZSTD_CCtx_params instances, we no longer need to redundantly track
individual compression parameters on our wrapper type.

This commit ports the C extension to the new world order.

This change was a bit invasive, as there were a few random
callers accessing the low-level struct members that we had
to delete.