compressionwriter: call close() from __exit__
This does two things. First, it allows us to reuse equivalent
logic to flush(FLUSH_FRAME). Second, it properly closes the
stream using the new close() API. Before, we just marked the
stream as closed without potentially taking appropriate action,
such as closing the underlying stream.

This change caused a lot of tests to fail because io.BytesIO
doesn't allow calling getvalue() after close(). So we had to
invent a wrapper type that works around this.