compressionwriter: introduce write_return_read argument
Our behavior of write() doesn't conform with the documented
behavior of io.RawIOBase. Specifically, we are returning the
number of bytes written to the underlying stream, not the
number of bytes read from the source.

We can't change the behavior without a backwards compatibility
break. So this commit introduces an argument to control the
return value of write().