cffi: consume remaining data in input buffer
Before, the CFFI implementation of ZstdDecompressionReader.read()
may skip input data left over in an internal buffer. This would
result in feeding incorrect input into the decompressor, which would
likely manifest as a malformed zstd data error.

As part of fixing this, we added a test to reproduce the failure.
And we also improved the fuzzing coverage of this method.

Closes #71.