decompressionreader: read across multiple frames
Previously, read() would stop reading at zstd frame boundaries.
After this commit, read() always attempts to read the number of
bytes requested, even if it means spanning zstd frames.

This change is backwards incompatible. But I think it is
the correct behavior, as it is what the `zstd` CLI tool does.

There is room to introduce a toggle for this behavior. I could
definitely see people wanting to force reads to emit data at
frame boundaries. We'll leave that for another commit.