ci: add Azure Pipelines configuration
Azure Pipelines supports Windows, Linux, and macOS. This commit
represents the start of our journey to consolidate CI onto Azure
Pipelines.

We implement support for building and testing on Windows, Linux,
and macOS. We build with both CPython and Anaconda (Windows only
for the moment). Wheels and Anaconda package are emitted as
artifacts. This likely get us parity with AppVeyor for Windows.