compressiondict: switch to fastcover for training
fastcover is the new training mechanism used by default. It appears the
old APIs are legacy and should be avoided.

This commit ports ``train_dictionary()`` to fastcover. As part of this,
we added arguments to control parameters exposed by the new mechanism.

Getting tests to pass was a pain. A problem was that different APIs have
different semantics. As part of the port, we consolidated onto calling
``ZDICT_optimizeTrainFromBuffer_fastCover()`` instead of 3 different
APIs. This ensures more consistency and is easier to debug. There will
likely be minor regressions in behavior as a result of this and other
changes. But the new state seems much better.

While this commit is annotated to KOLANICH, it was heavily refactored
by Gregory Szorc as part of incorporating the pull request. The CFFI
changes are all Gregory's. The original C code was highly refactored.