zstd: use single file zstandard library distribution
This keeps the distribution simpler, as fewer files are involved. This
also makes our custom build code simpler, as we no longer need to
annotate every file in the library distribution.

We couldn't remove zstd.h and zdict.h because CFFI isn't able to parse
the single file library. So we keep those around.

This change doesn't appear to impact compile times significantly.

This change may affect performance, as after this change we have a single
translation unit for the C extension (as opposed to dozens of object
files). This may enable some intra-translation unit optimizations within
the compiler similar to what we would have had if we used link time
optimization. Although I'm unsure of this, as I haven't benchmarked.