setup: initial support for building Rust extension
In order to build Rust extensions we need custom distutils classes
for invoking `cargo build`. This commit establishes that support.

To make hacking on the Rust extension more friendly, we add setup.py
arguments to control the building of the different backends.

This is all somewhat hacky at the moment. But it is better than nothing.

The distutils code was inspired by what I wrote for the
oxidized_importer package.