rust: implement compression parameters API
With this commit, test_data_structures.py passes using the Rust backend!

We had to switch to a Git revision of the zstd crates because the latest
released version doesn't properly support the `zstdmt` feature.

I'm not too crazy about the quality of the Rust code. It is mostly a
port of C semantics. We can refactor things later.