c-ext: use unified build for C extension sources
Fewer compiler invocations can reduce overhead and lead to more
compiler optimizations since all functions are in the same translation
unit.

But the real reason for this change is I want to switch to the single
file zstd distribution so things are simpler and presumably faster to
build. But this is blocked on
https://github.com/facebook/zstd/issues/2211.