zstandard: add typing stub file for zstandard
There are likely some subtle bugs in the type annotations. And we could
probably improve the type annotations by inventing new types for certain
constants. But this seems like a good start for the initial implementation.

While we don't yet have automated tests for typing, I believe my IDE says
the tests are good for the typing as defined in this commit. So I have
some confidence that we're on the right track here. In addition, mypy also
processes the zstandard package and tests without errors, which is a good
sign.

Closes #120.