compressionwriter: add closefd argument to control closing underlying stream
The existing behavior of automatically calling close() on the underlying
stream is not desired in all cases. See #76 for discussion.

This commit adds an argument to ZstdCompressionWriter to control whether
the inner stream is closed when the outer stream is closed. This will give
consumers control over how to behave.

We plan to implement this feature for all file-like types. After we're
done, we'll revisit the default behavior.