c-ext: properly check for errors after PyObject_CallMethod()
Before this commit, we didn't check for errors after various instances
of PyObject_CallMethod(). This resulted in a run-time error because an
exception was set but the function didn't return NULL.

I audited all occurrences where we call Python functions and made sure
to handle errors in every case.

Tests were added to ensure behavior is as expected. In one test, behavior
between the C and CFFI backends differed. So I changed the behavior of the
C backend. This pattern of recasting the exception occurs elsewhere and
likely only occurs in the C backend. We'll probably want to add tests and
clean up this behavior as well.

Closes #102.