setup: change distribution metadata so it is consistent
Distribution metadata should be constant so it doesn't vary based on
how the distribution was built. This commit changes it so it is so.

We list cffi as a required dependency on PyPy and an optional dependency
everywhere else.

I'm unsure if the new metadata is the best end state. But it seems
strictly better than before. In absence of feedback that this doesn't
work, I'm going to say closes #97.