docs: use docstrings for compressor API documentation
Previously, the state of our API docs was a bit lacking. We had docstrings
defined in the C backend. And we had unstructured rst docs in the Sphinx
documentation.

This commit attempts to modernize things a bit by leveraging Sphinx's
support for parsing docstrings to generate the API documentation.

Because maintaining docs in C code is a bit cumbersome compared to
Python, we move the docstrings to the CFFI backend, where they can be
defined in Python source code. A side-effect of this is that the C
backend no longer has meaningful values for ``__doc__``. This is
unfortunate. But I think it is the lesser evil of having to maintain
multiple copies of the docs. It's worth noting that you can't dynamically
update ``__doc__`` of a type backed by C. Otherwise we could do this in
``__init__.py`` when loading the C backend.

As part of this, we also merge the old Sphinx rst docs into the
docstrings. Much of this content should belong in docstrings of
appropriate types. But this will require additional refactoring. I
thought a good milestone would be to port the prior Sphinx docs to
docstrings then clean things up later.