c-ext: conditionally enable features relying on pool APIs
The pool APIs are not exported from the library nor are they present
in public headers.

This commit changes behavior of the C backend to conditionally enable
features relying on these symbols. The practical effect is that these
features will only be available if building against the single file /
bundled libzstd and will no longer be available when building against a
system libzstd.

This effectively closes #106.

A side-effect of this change is that --system-zstd likely regressed
due to inability to find zstd headers. We'll need to add a setup.py
argument to plumb the header search path through. But this was a
pre-existing condition.