rust: emit partial output chunk
This matches the behavior of the other backends and fixes a test failure.
As part of this, we also set finished_output when decompression is done.

With this change, all non-fuzzing tests now pass! Unfortunately, some
fuzzing tests still failure, which points to the existence of subtle bugs
in the Rust backend.