tests: optimize hypothesis interaction
Calling `strategies.sampled_from()` inside the function is expensive
because constructing the return value is expensive because it calls
`repr()` and our inputs are very large. This leads to some very memory
I/O intensive operations inside the Python interpreter.

On my machine, this reduces the run time in CI profile mode from ~43s
to ~2.5s. Not bad! Results in CI should be even more profound.