cffi: Use the distutils preprocessor when available
On Unix like systems, distutils will usually set the preprocessor
attribute to the appropriate command. On MSVC it never sets the
preprocessor, though. In either case, try to use it and only fallback to
using the compiler if needed. Using compiler.compiler[0] breaks use of
ccache when inserted in the CC environment variable.

This is a regression from 0d1329ab426dde7dc319905da0656cd44db41988.

Closes #178.