tests: add test for calling decompress multiple times
This test was previously failing in the C backend due to the
`assert(!self->unused_data)`. The previous commit refactored the
code so `self->unused_data` is only assigned after zstd frame
completion (not exhaustion of input). So the assertion should be
upheld.

The bug fixes are also documented in the changelog.

Closes #180.