global: remove use of deprecated `ZSTD_copyDCtx()`
It was deprecated in 1.5.4. So we started seeing compiler warnings
post upgrade.

The C change was pretty straightforward. The Rust change not so much,
as we had to teach Rust to load the dictionary on the created dctx.
Since Rust didn't have this code before, it appears this was actually
optional as `ZSTD_copyDCtx()` had the same effect of copying any
referenced dictionary data?