rust-ext: remove use of `#[pyproto]`
It is deprecated in PyO3 0.16.

This was pretty straightforward except for `buffers.rs`, where
`Py_buffer` handling is a bit wonky. It might be possible to coerce
the signature into taking a `PyBuffer<u8>`. But I couldn't figure out
the magic incantation. All we do is coerce the argument to a pointer
anyway and the argument is constructed by Python to populate. So I'm
pretty sure we don't need to be concerned with type confusion.