decompressionobj: implement reading across frames
All APIs need to gain the ability to transparently read across
frames. This commit teaches the stdlib decompressionobj interface to
transparently read across frames.

Behavior is controlled via a boolean named argument. It defaults to
False to preserve existing behavior.

Related to #196.

I haven't tested this very thoroughly. But the added fuzzing test
not failing is usually a good indicator that this works as intended.