ci: add ARM macOS test workers
Let's stand up CI coverage for ARM macOS now that the GitHub
Actions runners are available for free.

I couldn't get the Rust backend to build trivially. Looks like it
is due to the CI workers installing a universal binary for Python
and this not playing well with the Rust build backend. Will need
more work here. For another commit.