zstd: cherry-pick qsort commits
This change reflects cherry-picking of the following upstream
commits from facebook/zstd#4312 and facebook/zstd#4336:

* ebfa660b8297170fb0742ac8959aa64940292853
* 51b6e79f659f7ba18bf7c4bd2679ff58711aa34d
* dcf675886bc138c03c3f157f7464798c4140829f
* a9b8fef2e808c227058da4d0f41768c877cfba4c
* 4d53e27144b5ad02c6c37c27a998300986a4adaa

We need these commits so we can force musl <1.2 to use an
appropriate qsort implementation. And the commits seem to shore
up the feature detection and squash some bugs. So I'm fine carrying
them before we would otherwise get them in 1.5.8.