ext: declare GIL as not used
Without this, the GIL is enabled when loading the extension module since we
don't declare nogil compatibility.

Let's allow loading without the GIL so we can start meaningfully testing
running without the GIL.

We still don't formally support free-threaded interpreters. But we do
introduce the trove classifier at level 1 to indicate the feature is
available.