ci: use uv + cibuildwheel to build all wheels
Previously, wheel builds were a combination of using `cibuildwheel`
and `pip wheel`. On Linux, we manually spawned a container image and
provided our own script for building the wheel.

This commit standardizes on using `cibuildwheel` everywhere.

We run `cibuildwheel` via `uv` as a dependency package so behavior
is more deterministic. Except for the musllinux emulated builds, where
uv isn't available.