build: adopt `manylinux_2_28` on 3.14
This is only needed for ppc64le, which breaks when we introduce cffi
2.0 for free-threaded support (it appears the pre-built 2.0.0b wheels
have an glibc 2.22 symbol - unsure if this is a bug).

But, `manylinux2014` is getting long in the tooth. I'm not convinced we
need to provide binaries that work on distros from 2013. Especially for
a Python version released in 2025. So let's just bump the ABI on all
3.14 wheels.