cffi: Move CFFI logic from global scope into `get_ffi()` function
Move the CFFI-related logic from the global scope of `make_cffi.py`
file to a `get_ffi()` function.  This makes it possible to import
the file without executing the code immediately, and it will make it
possible to parametrize the invocation using explicit function
parameters.  Along with the change, I've renamed the variables to
lowercase since they are no longer global.

Signed-off-by: Michał Górny <mgorny@gentoo.org>